refactor: extract the bytea tuplesort wrapper from keyset.rs
Move the Tuplesortstate-over-bytea helper out of keyset.rs into
postgres/tuplesort.rs so the partitioned index build can spill its
(pid, ctid) records through the same wrapper. Behavior preserved; the
pg15 getdatum copy is freed by the caller as before.
