+++ pg_search/src/postgres/mod.rs
@@ -66,6 +66,7 @@ pub mod pdb_owned_value;
 pub mod planner_warnings;
 pub mod rel;
 pub mod storage;
+pub mod tuplesort;
 pub mod types;
 pub mod types_arrow;
 pub mod utils;
